Window creation reimplementation
Previous approach was inflexible in how a client can create windows - the original idea was to have a manager "know" set of windows it can create. This creates an issue when it's not possible to guess what kinds of windows will be drawn. In hindsight, it was an unneccessary restriction, that gave no pros, but had obvious design flaws (how can client update window's props? etc)